style: decoupling `kipdb::TransactionIter`
