+++ src/execution/volcano/dql/join/nested_loop_join.rs
@@ -328,7 +328,7 @@ mod test {
     use crate::expression::ScalarExpression;
     use crate::planner::operator::values::ValuesOperator;
     use crate::planner::operator::Operator;
-    use crate::storage::kip::KipStorage;
+    use crate::storage::kipdb::KipStorage;
     use crate::storage::Storage;
     use crate::types::value::DataValue;
     use crate::types::LogicalType;
