+++ src/optimizer/core/statistics_meta.rs
@@ -3,8 +3,7 @@ use crate::errors::DatabaseError;
 use crate::expression::range_detacher::Range;
 use crate::optimizer::core::cm_sketch::CountMinSketch;
 use crate::optimizer::core::histogram::Histogram;
-use crate::storage::kip::StatisticsMetaCache;
-use crate::storage::Transaction;
+use crate::storage::{StatisticsMetaCache, Transaction};
 use crate::types::index::IndexId;
 use crate::types::value::DataValue;
 use serde::{Deserialize, Serialize};
