+++ tests/sqllogictest/src/lib.rs
@@ -1,6 +1,6 @@
 use fnck_sql::db::Database;
 use fnck_sql::errors::DatabaseError;
-use fnck_sql::storage::kip::KipStorage;
+use fnck_sql::storage::kipdb::KipStorage;
 use sqllogictest::{AsyncDB, DBOutput, DefaultColumnType};
 use std::time::Instant;
